Split CC index filegroup helpers
